lexer: Replace old lexer with yacc-style algorithm
This CL is essentially a full rewrite of the tokenizer to make it more
maintainable as the project scales. The new algorithm is basically lex.

The old algorithm was "smarter" in the sense that the tokenizer had more
grammar awareness and could use that to write more informed token names,
but this added intelligence came at the cost of unmaintainable complexity.
The tokenizer was considerably more complicated than the parser logic which
felt backwards.